+++ kernel/vmem/paging.rs
@@ -1,5 +1,4 @@
-
 use pache::{KiB, MiB};
 
-pub const PAGE_SIZE: usize = 4 * KiB;
-pub const LPAGE_SIZE: usize = 2 * MiB;
+pub const PAGE_SIZE : u64 = 4 * KiB;
+pub const LPAGE_SIZE: u64 = 2 * MiB;
