+++ patchouli/src/units.rs
@@ -1,6 +1,6 @@
 #![allow(non_upper_case_globals)]
 
-pub const KiB : usize = 1024;
-pub const MiB : usize = KiB * KiB;
-pub const GiB : usize = KiB * MiB;
-pub const TiB : usize = KiB * GiB;
+pub const KiB : u64 = 1024;
+pub const MiB : u64 = KiB * KiB;
+pub const GiB : u64 = KiB * MiB;
+pub const TiB : u64 = KiB * GiB;
